Add unfriend functionality to remove users from friend list

+++ src/friends/services.py
@@ -8,27 +8,31 @@ User = get_user_model()
 
 
 class FriendService:
-    def accept_friend_request(self, from_user, to_user):
+    def accept_friend_request(self, to_user, from_user):
+        """
+        to_user — тот, кто принимает (request.user),
+        from_user — тот, кто отправил заявку.
+        """
         try:
             friend_request = FriendRequestModel.objects.get(from_user=from_user, to_user=to_user)
-
-            friendship = FriendShipModel(user1=from_user, user2=to_user)
-            friendship.save()
-
-            friend_request.delete()
-
         except FriendRequestModel.DoesNotExist:
             raise ValidationError("Friend request not found")
 
+        # Сортировка пользователей по id
+        user1, user2 = sorted([to_user, from_user], key=lambda u: u.id)
 
+        if FriendShipModel.objects.filter(user1=user1, user2=user2).exists():
+            raise ValidationError("Already friends")
+
+        FriendShipModel.objects.create(user1=user1, user2=user2)
+        friend_request.delete()
 
     def send_friend_request(self, from_user, to_user):
         if from_user == to_user:
             raise ValueError("Нельзя добавить себя в друзья")
 
         existing = FriendRequestModel.objects.filter(
-            (Q(from_user=from_user) & Q(to_user=to_user)) |
-            (Q(from_user=to_user) & Q(to_user=from_user))
+            (Q(from_user=from_user) & Q(to_user=to_user)) | (Q(from_user=to_user) & Q(to_user=from_user))
         ).exists()
         if existing:
             raise ValueError("Заявка уже существует в любом направлении")
@@ -43,21 +47,14 @@ class FriendService:
         except FriendRequestModel.DoesNotExist:
             raise ValueError("Заявка на дружбу не найдена или уже отменена")
 
-
-
-
-
-
-    def reject_request(self, from_user, to_user):...
+    def reject_request(self, from_user, to_user): ...
 
     # Отклоняем заявку
 
-    def remove_friend(self, user1, user2):...
+    def remove_friend(self, user1, user2): ...
 
     # Удаляем Friendship
 
 
 class BlockService:
     pass
-
-
